For #38944, Separated prepare from init and launch app logic
